@@ -4,9 +4,10 @@ import 'package:dicoding_news_app/ui/article_detail_page.dart';
 import 'package:dicoding_news_app/widgets/platform_widget.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/widgets.dart';
 
 class ArticleListPage extends StatelessWidget {
+  const ArticleListPage({Key? key}) : super(key: key);
+
   Widget _buildList(BuildContext context) {
     return FutureBuilder<String>(
       future: DefaultAssetBundle.of(context).loadString('assets/articles.json'),
@@ -50,7 +51,7 @@ class ArticleListPage extends StatelessWidget {
   Widget _buildAndroid(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text('News App'),
+        title: const Text('News App'),
       ),
       body: _buildList(context),
     );
@@ -58,7 +59,7 @@ class ArticleListPage extends StatelessWidget {
 
   Widget _buildIos(BuildContext context) {
     return CupertinoPageScaffold(
-      navigationBar: CupertinoNavigationBar(
+      navigationBar: const CupertinoNavigationBar(
         middle: Text('News App'),
         transitionBetweenRoutes: false,
       ),
