@@ -7,7 +7,7 @@ class ArticleWebView extends StatelessWidget {
 
   final String url;
 
-  const ArticleWebView({required this.url});
+  const ArticleWebView({Key? key, required this.url}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
