@@ -1,5 +1,6 @@
 import 'dart:io';
 
+import 'package:dicoding_news_app/common/styles.dart';
 import 'package:dicoding_news_app/ui/article_list_page.dart';
 import 'package:dicoding_news_app/ui/settings_page.dart';
 import 'package:dicoding_news_app/widgets/platform_widget.dart';
@@ -9,20 +10,22 @@ import 'package:flutter/material.dart';
 class HomePage extends StatefulWidget {
   static const routeName = '/home_page';
 
+  const HomePage({Key? key}) : super(key: key);
+
   @override
-  _HomePageState createState() => _HomePageState();
+  State<HomePage> createState() => _HomePageState();
 }
 
 class _HomePageState extends State<HomePage> {
   int _bottomNavIndex = 0;
   static const String _headlineText = 'Headline';
 
-  List<Widget> _listWidget = [
+  final List<Widget> _listWidget = const [
     ArticleListPage(),
     SettingsPage(),
   ];
 
-  List<BottomNavigationBarItem> _bottomNavBarItems = [
+  final List<BottomNavigationBarItem> _bottomNavBarItems = [
     BottomNavigationBarItem(
       icon: Icon(Platform.isIOS ? CupertinoIcons.news : Icons.public),
       label: _headlineText,
@@ -43,6 +46,7 @@ class _HomePageState extends State<HomePage> {
     return Scaffold(
       body: _listWidget[_bottomNavIndex],
       bottomNavigationBar: BottomNavigationBar(
+        selectedItemColor: secondaryColor,
         currentIndex: _bottomNavIndex,
         items: _bottomNavBarItems,
         onTap: _onBottomNavTapped,
@@ -52,7 +56,10 @@ class _HomePageState extends State<HomePage> {
 
   Widget _buildIos(BuildContext context) {
     return CupertinoTabScaffold(
-      tabBar: CupertinoTabBar(items: _bottomNavBarItems),
+      tabBar: CupertinoTabBar(
+        items: _bottomNavBarItems,
+        activeColor: secondaryColor,
+      ),
       tabBuilder: (context, index) {
         return _listWidget[index];
       },
