@@ -6,10 +6,12 @@ import 'package:flutter/material.dart';
 class SettingsPage extends StatelessWidget {
   static const String settingsTitle = 'Settings';
 
+  const SettingsPage({Key? key}) : super(key: key);
+
   Widget _buildAndroid(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text(settingsTitle),
+        title: const Text(settingsTitle),
       ),
       body: _buildList(context),
     );
@@ -17,7 +19,7 @@ class SettingsPage extends StatelessWidget {
 
   Widget _buildIos(BuildContext context) {
     return CupertinoPageScaffold(
-      navigationBar: CupertinoNavigationBar(
+      navigationBar: const CupertinoNavigationBar(
         middle: Text(settingsTitle),
       ),
       child: _buildList(context),
@@ -29,7 +31,7 @@ class SettingsPage extends StatelessWidget {
       children: [
         Material(
           child: ListTile(
-            title: Text('Dark Theme'),
+            title: const Text('Dark Theme'),
             trailing: Switch.adaptive(
               value: false,
               onChanged: (value) {
@@ -39,11 +41,11 @@ class SettingsPage extends StatelessWidget {
                         barrierDismissible: true,
                         builder: (context) {
                           return CupertinoAlertDialog(
-                            title: Text('Coming Soon!'),
-                            content: Text('This feature will be coming soon!'),
+                            title: const Text('Coming Soon!'),
+                            content: const Text('This feature will be coming soon!'),
                             actions: [
                               CupertinoDialogAction(
-                                child: Text('Ok'),
+                                child: const Text('Ok'),
                                 onPressed: () {
                                   Navigator.pop(context);
                                 },
@@ -56,14 +58,14 @@ class SettingsPage extends StatelessWidget {
                         context: context,
                         builder: (context) {
                           return AlertDialog(
-                            title: Text('Coming Soon!'),
-                            content: Text('This feature will be coming soon!'),
+                            title: const Text('Coming Soon!'),
+                            content: const Text('This feature will be coming soon!'),
                             actions: [
                               TextButton(
                                 onPressed: () {
                                   Navigator.pop(context);
                                 },
-                                child: Text('Ok'),
+                                child: const Text('Ok'),
                               ),
                             ],
                           );
