@@ -5,7 +5,7 @@ class PlatformWidget extends StatelessWidget {
   final WidgetBuilder androidBuilder;
   final WidgetBuilder iosBuilder;
 
-  PlatformWidget({required this.androidBuilder, required this.iosBuilder});
+  const PlatformWidget({Key? key, required this.androidBuilder, required this.iosBuilder}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
